Refactor: mongoDB 폴더 명시 (#313)
