+++ src/main/java/com/partnerd/service/kafkaService/KafkaConsumer.java
@@ -3,14 +3,12 @@ package com.partnerd.service.kafkaService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.partnerd.converter.ChatConverter;
 import com.partnerd.domain.chat.ChatMessage;
-import com.partnerd.repository.chatRoomRepository.ChatMessageRepository;
+import com.partnerd.repository.chatRoomRepository.mongoRepository.ChatMessageRepository;
 import com.partnerd.web.dto.chatDTO.ChatDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.kafka.annotation.KafkaListener;
-import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Service;
 import java.time.Instant;
